core: rename custom.js
This avoids import conflicts with the custom.js file used by the s
atskeeper.com webapp
