+++ configurator.html
@@ -292,5 +292,5 @@
   <script type="text/javascript" src="/static/js/jsonEditor.js" defer></script>
   <script type="text/javascript" src="/static/js/FileSaver.js" defer></script>
   <script type="text/javascript" src="/static/js/cryptojs.js" defer></script>
-  <script src="/static/js/custom.js" type="module" defer></script>
+  <script src="/static/js/editor.js" type="module" defer></script>
 </html>
