+++ viewer.html
@@ -89,5 +89,5 @@
   <script type="text/javascript" src="/static/js/jsonEditor.js" defer></script>
   <script type="text/javascript" src="/static/js/FileSaver.js" defer></script>
   <script type="text/javascript" src="/static/js/cryptojs.js" defer></script>
-  <script type="module" src="/static/js/custom.js" defer></script>
+  <script type="module" src="/static/js/editor.js" defer></script>
 </html>
